heavy-agentkit-bridge/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "heavy-agentkit-bridge"
version = "0.1.0"
description = "Bridge service connecting Heavy with AgentKit and ElizaOS"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "Proprietary" }
authors = [
    { name = "Heavy Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "websockets>=10.1",
    "aiohttp>=3.8.1",
    "pydantic>=1.8.2",
    "pyyaml>=6.0",
    "cryptography>=36.0.0",
    "python-dotenv>=0.19.2",
    "coinbase-agentkit>=0.1.0",
    "prometheus-client>=0.14.1",
    "aiodns>=3.0.0",        # For async DNS resolution
    "cchardet>=2.1.7",      # For faster encoding detection
    "ujson>=5.4.0",         # For faster JSON processing
    "uvloop>=0.16.0",       # For faster event loop
    "systemd-python>=234",  # For systemd integration (Linux only, optional)
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "pytest-asyncio>=0.16.0",
    "black>=22.1.0",
    "mypy>=0.910",
    "flake8>=4.0.1",
    "isort>=5.10.1",
    "coverage>=6.3.2",
    "pytest-cov>=3.0.0",
]
monitoring = [
    "prometheus-client>=0.14.1",
    "aiohttp>=3.8.1",
]
security = [
    "cryptography>=36.0.0",
    "cffi>=1.15.0",
    "bcrypt>=3.2.0",
]
docs = [
    "mkdocs>=1.3.0",
    "mkdocs-material>=8.1.11",
    "mkdocstrings>=0.18.1",
]

[project.urls]
"Homepage" = "https://github.com/your-org/heavy-agentkit-bridge"
"Bug Tracker" = "https://github.com/your-org/heavy-agentkit-bridge/issues"

[project.scripts]
heavy-bridge = "src.main:cli_main"
heavy-agent = "src.elizaos.elizaos_agent:cli_main"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
